+++ client/src/components/images/CharacterImg.js
@@ -2,14 +2,18 @@ import React from "react";
 import PropTypes from "prop-types";
 import { useSelector } from "react-redux";
 
-import { getImgName } from "../../utils/utils";
-
 const CharacterImg = ({ name, imgClass, ...otherProps }) => {
     const { nightMode } = useSelector((state) => state.appState);
 
+    const formattedName = name
+        .replaceAll(" ", "_")
+        .replaceAll('"', "")
+        .replaceAll(".", "")
+        .toLowerCase();
+
     return (
         <img
-            src={`/img/characters/${getImgName(name)}.png`}
+            src={`/img/webp/characters/C-${formattedName}.webp`}
             className={[
                 imgClass || "char-img",
                 nightMode ? "nightMode-img-bg" : null